complete gitlab

+++ bin/boot_docker.sh
@@ -1,16 +1,58 @@
 source setting.conf
 
+##### goto target dir #####
+target_list=$(ls ../conf)
+echo "------- choose target --------"
+index=-1
+for target in ${target_list[@]}
+do
+  index=$((index+1))
+  echo "    [$index]  $target"
+done
+echo "---------------"
+
+#docker ps|awk '{print $NF}'
+echo "which target you want enter[0-$index]："
+read target_index
+echo "index: $target_index"
+
+#verify target_index
+is_exist=0
+
 #goto target dir
-sh go_target.sh $@
-exit $?
+index=-1
+for target in ${target_list[@]}
+do
+  index=$((index+1))
+  if [ $index = "$target_index" ]; then
+    is_exist=1
+	cd ../conf/$target
+	echo "current path is `pwd`"
+  fi
+done
 
-#Fetch name of container and image from $param_file
-sh fetch_docker_info.sh
+#verify fasle,then exit 1
+if [[ $is_exist -ne 1 ]];then
+    echo "Error for target_index : $target_index"
+    exit 1
+fi
 
+#Fetch name of container and image
+export IMAGE_NAME=`cat $DOCKER_INFO_FILE | grep IMAGE_NAME | tail -1 | awk -F '=' '{print $2}'`
+echo "IMAGE_NAME is $IMAGE_NAME"
+export CONTAINER_NAME=`cat $DOCKER_INFO_FILE | grep CONTAINER_NAME | tail -1 | awk -F '=' '{print $2}'`
+echo "CONTAINER_NAME is $CONTAINER_NAME"
+
+#log
+Date=`date "+%Y-%m-%d %H:%M:%S"`
+echo "IMAGE_NAME is $IMAGE_NAME"
+echo "$Date IMAGE_NAME=$IMAGE_NAME" >> ${LOG_FILE}
+
+##### start to boot #####
 docker pull ${IMAGE_NAME}
 docker rm -f $CONTAINER_NAME
 
-sh docker_run.sh
+sh $DOCKER_RUN_FILE
 
 if [[ $? != 0 ]];then
     echo "Fail to boot image $IMAGE_NAME";
